Done demo.

+++ server/src/main/java/services/impl/EventServiceImpl.java
@@ -2,6 +2,7 @@ package services.impl;
 
 import entity.Event;
 import entity.RepeatType;
+import entity.User;
 import exception.EventNotFound;
 import exception.UserNotFound;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,9 +15,7 @@ import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -25,14 +24,11 @@ public class EventServiceImpl implements EventService {
     private EventRepository eventRepository;
     @Autowired
     private UserRepository userRepository;
-    @Override
-    public Event addEvent(Event event) {
-        return eventRepository.save(event);
-    }
+
 
     @Override
-    public Event addEvent(long userId, String summary, LocalDateTime start, Duration duration, RepeatType repeatType) throws UserNotFound {
-        var _user = userRepository.findById(userId).orElseThrow(() -> new UserNotFound(userId));
+    public Event addEvent(User user, String summary, LocalDateTime start, Duration duration, RepeatType repeatType) throws UserNotFound {
+        var _user = userRepository.findById(user.getId()).orElseThrow(() -> new UserNotFound(user.getId()));
         var newEvent = new Event();
         newEvent.setDuration(duration);
         newEvent.setSummary(summary);
@@ -50,55 +46,51 @@ public class EventServiceImpl implements EventService {
 
     @Override
     public List<Event> getEvents(long userId) throws UserNotFound {
-        return eventRepository.findEvents(userRepository.getReferenceById(userId));
+        return eventRepository.findEvents(userRepository.findById(userId).orElseThrow(() -> new UserNotFound(userId)));
     }
 
     private boolean onThisDay(Event e, LocalDate date){
-        var start = e.getStart();
-
-        var dtf = switch (e.getRepeat()){
-            case NONE,HOURLY -> DateTimeFormatter.ofPattern("yyyy-MM-dd");
-            case DAILY -> DateTimeFormatter.ofPattern("yyyy-MM");
-            case WEEKLY -> DateTimeFormatter.ofPattern("F");
-            case MONTHLY -> DateTimeFormatter.ofPattern("yyyy-dd");
+        var start = e.getStart().toLocalDate();
+        return switch (e.getRepeat()){
+            case NONE ->  start.equals(date);
+            case HOURLY, DAILY -> true;
+            case WEEKLY -> start.getDayOfWeek().equals(date.getDayOfWeek());
+            case MONTHLY -> start.getDayOfMonth() == date.getDayOfMonth();
         };
-        return start.format(dtf).equals(date.format(dtf));
     }
 
     private boolean atThisTime(Event e, LocalTime time){
-        var start = e.getStart();
-        var dtf = switch (e.getRepeat()){
-            case HOURLY -> DateTimeFormatter.ofPattern("HH");
-            default -> DateTimeFormatter.ofPattern("HH:mm");
+        var start = e.getStart().toLocalTime();
+        return switch (e.getRepeat()){
+            case HOURLY -> start.getMinute() == time.getMinute();
+            default -> start.equals(time);
         };
-        return start.format(dtf).equals(time.format(dtf));
     }
 
     @Override
     public List<Event> getEventsByDate(long userId, LocalDate date) throws UserNotFound {
         List<Event> allEvents = getEvents(userId);
         return allEvents
-                .parallelStream()
-                .map(event -> onThisDay(event, date)?event:null)
-                .filter(Objects::nonNull).collect(Collectors.toList());
+                .stream()
+                .filter(event -> onThisDay(event, date)).collect(Collectors.toList());
     }
 
     @Override
     public List<Event> getEventsByTime(long userId, LocalTime time) throws UserNotFound {
         List<Event> allEvents = getEvents(userId);
         return allEvents
-                .parallelStream()
-                .map(event -> atThisTime(event, time)?event:null)
-                .filter(Objects::nonNull).collect(Collectors.toList());
+                .stream()
+                .filter(event -> atThisTime(event, time)).collect(Collectors.toList());
     }
 
     @Override
     public List<Event> getEventsByDateTime(long userId, LocalDate date, LocalTime time) throws UserNotFound {
-        return getEventsByDate(userId,date).parallelStream()
-                .map(event -> atThisTime(event, time)?event:null)
-                .filter(Objects::nonNull).collect(Collectors.toList());
+        return getEventsByDate(userId,date).stream()
+                .filter(event -> atThisTime(event, time))
+                .collect(Collectors.toList());
     }
 
+    @Deprecated
     @Override
     public List<Event> getEventsInRange(long userId, LocalDateTime start, LocalDateTime end) throws UserNotFound {
         if (!userRepository.existsById(userId)) throw new UserNotFound(userId);
@@ -106,14 +98,71 @@ public class EventServiceImpl implements EventService {
     }
 
     @Override
-    public void deleteEvent(long eventId) {
+    public void deleteEvent(long eventId) throws EventNotFound {
+        if (!eventRepository.existsById(eventId))
+            throw new EventNotFound(eventId);
         eventRepository.deleteById(eventId);
     }
 
+
+//
+//    @Override
+//    public Event updateSummary(long eventId, String summary) throws EventNotFound {
+//        var currentEvent = eventRepository.findById(eventId).orElseThrow(()->new EventNotFound(eventId));
+//        eventRepository.deleteById(eventId);
+//        currentEvent.setSummary(summary);
+//        return eventRepository.save(currentEvent);
+//    }
+//
+//    @Override
+//    public Event updateStart(long eventId, LocalDateTime start) throws EventNotFound {
+//        var currentEvent = eventRepository.findById(eventId).orElseThrow(()->new EventNotFound(eventId));
+//        eventRepository.deleteById(eventId);
+//        currentEvent.setStart(start);
+//        return eventRepository.save(currentEvent);
+//    }
+//    @Override
+//    public Event updateDuration(long eventId, Duration duration) throws EventNotFound {
+//        var currentEvent = eventRepository.findById(eventId).orElseThrow(()->new EventNotFound(eventId));
+//        eventRepository.deleteById(eventId);
+//        currentEvent.setDuration(duration);
+//        return eventRepository.save(currentEvent);
+//    }
+//    @Override
+//    public Event updateRepeat(long eventId, RepeatType repeatType) throws EventNotFound {
+//        var currentEvent = eventRepository.findById(eventId).orElseThrow(()->new EventNotFound(eventId));
+//        eventRepository.deleteById(eventId);
+//        currentEvent.setRepeat(repeatType);
+//        return eventRepository.save(currentEvent);
+//    }
     @Override
-    public Event updateEvent(long oldEventId, Event newEvent) {
-        eventRepository.deleteById(oldEventId);
-        return eventRepository.save(newEvent);
+    public Event updateSummary(long eventId, String summary) throws EventNotFound {
+        if (!eventRepository.existsById(eventId))
+            throw new EventNotFound(eventId);
+        eventRepository.updateSummary(eventId, summary);
+        return eventRepository.findById(eventId).orElseThrow(() -> new EventNotFound(eventId));
+    }
+
+    @Override
+    public Event updateStart(long eventId, LocalDateTime start) throws EventNotFound {
+        if (!eventRepository.existsById(eventId))
+            throw new EventNotFound(eventId);
+        eventRepository.updateStart(eventId, start);
+        return eventRepository.findById(eventId).orElseThrow(() -> new EventNotFound(eventId));
+    }
+    @Override
+    public Event updateDuration(long eventId, Duration duration) throws EventNotFound {
+        if (!eventRepository.existsById(eventId))
+            throw new EventNotFound(eventId);
+        eventRepository.updateDuration(eventId, duration);
+        return eventRepository.findById(eventId).orElseThrow(() -> new EventNotFound(eventId));
+    }
+    @Override
+    public Event updateRepeat(long eventId, RepeatType repeatType) throws EventNotFound {
+        if (!eventRepository.existsById(eventId))
+            throw new EventNotFound(eventId);
+        eventRepository.updateRepeat(eventId, repeatType);
+        return eventRepository.findById(eventId).orElseThrow(() -> new EventNotFound(eventId));
     }
 
 }
